+++ app/bike_model.py
@@ -26,6 +26,7 @@ class CatalogFamily:
 @dataclass
 class SubscriptionBikeFamily:
     """type of subscription with information about it."""
+
     subscribe_id: int
     chat_id: int
     bike_family: str
