+++ app/bot.py
@@ -9,17 +9,12 @@ from typing import Generator, List
 
 from aiogram import Bot, Dispatcher, executor, types
 from aiogram.contrib.fsm_storage.memory import MemoryStorage
-from aiogram.dispatcher import FSMContext
-from aiogram.dispatcher.filters.state import StatesGroup, State
 from aiogram.utils.markdown import hlink
 
-from app.bike_model import Bike, CatalogFamily, SubscriptionBikeFamily
+from app import bot_subscription
+from app.bike_model import Bike, CatalogFamily
 from app.settings import app_settings
-from app.storage import get_catalog, create_subscription
-
-
-class CreateSubscription(StatesGroup):
-    family_name = State()
+from app.storage import get_catalog
 
 
 async def send_welcome(message: types.Message) -> None:
@@ -29,8 +24,8 @@ async def send_welcome(message: types.Message) -> None:
         'I will show you which canyon bicycles are available in the store.',
         '/catalog - to see all catalog.',
         '/subscribe - to get the message, when the bike family you want in the stock.',
-#        '/unsubscribe - not to receive messages about bike family.',
-#        '/subscriptions_list - check if you waiting for any messages.',
+        # '/unsubscribe - not to receive messages about bike family.',
+        # '/subscriptions_list - check if you waiting for any messages.',
     ))
 
     await message.answer(answer_text)
@@ -69,52 +64,6 @@ async def show_catalog(message: types.Message) -> None:  # noqa: WPS210
         )
 
 
-async def start_subscription(message: types.Message) -> None:
-    """Ask a bike family name."""
-    reply_text = '\n'.join((
-        'Please, write the bike family name.When it will be available we will let you know!',
-        '/cancel - to cancel the action.',
-    ))
-    await CreateSubscription.family_name.set()
-    await message.reply(reply_text)
-
-
-async def cancel_subscription(message: types.Message, state: FSMContext):
-    """Allow user to cancel action via /cancel command"""
-
-    current_state = await state.get_state()
-    if current_state is None:
-        return
-
-    await state.finish()
-    await message.reply('Cancelled.')
-
-
-async def process_subscription(message: types.Message, state: FSMContext) -> None:
-    """Create the subscription."""
-
-    created_subscription: SubscriptionBikeFamily = await create_subscription(message.chat.id, message.text)
-
-    await state.finish()
-    await message.reply(f'Got it! When "{created_subscription.bike_family}" will be available we will let you know!')
-
-
-# Todo может и не пригодится
-async def remove_subscription(message: types.Message) -> None:
-    """Ask a bike family name and delete it from subscription list."""
-    answer_text = '\n'.join((
-        'Please, write the bike family name you dont want to have messages about.',
-    ))
-
-    await message.answer(answer_text)
-
-
-async def show_subscriptions(message: types.Message) -> None:
-    """Shows all subscriptions."""
-    # Todo вывереси список подписок (посмотреть на списке байков)
-    ...
-
-
 def main() -> None:
     """Telegram bot app runner."""
     bot = Bot(token=app_settings.bot_token)
@@ -123,11 +72,14 @@ def main() -> None:
     router = Dispatcher(bot, storage=storage)
     router.register_message_handler(send_welcome, commands=['start', 'help'])
     router.register_message_handler(show_catalog, commands=['catalog'])
-    router.register_message_handler(start_subscription, commands=['subscribe'])
-    router.register_message_handler(cancel_subscription, state='*', commands=['cancel'])
-    router.register_message_handler(process_subscription, state=CreateSubscription.family_name)
-    router.register_message_handler(remove_subscription, commands=['unsubscribe'])
-    router.register_message_handler(show_subscriptions, commands=['subscriptions_list'])
+    router.register_message_handler(bot_subscription.start_subscription, commands=['subscribe'])
+    router.register_message_handler(bot_subscription.cancel_subscription, state='*', commands=['cancel'])
+    router.register_message_handler(
+        bot_subscription.process_subscription,
+        state=bot_subscription.CreateSubscription.family_name,
+    )
+    router.register_message_handler(bot_subscription.remove_subscription, commands=['unsubscribe'])
+    router.register_message_handler(bot_subscription.show_subscriptions, commands=['subscriptions_list'])
     executor.start_polling(router, skip_updates=True)
